Online incremental memory testing.
Sometimes Redis crashes reported by users are caused by corrupted
memory in the hardware where Redis runs.

This has a big cost for the Redis communtiy:

1) Possible data corruption.
2) Serives unavailability.
3) Wasted Redis developers and Redis users debugging time.

For this reason at some time Redis introduced a way to test memory,
running redis-server --test-memory. However while this turned out to be
extremely useful as users now have a simple way to test the computer
memory without the need to reboot or install other software, not
everybody runs this test, or run it for enough time to detect errors
before a fresh installation of Redis.

An alternative approach is to also test memory at runtime incrementally,
and report memory errors in the INFO output, that is also included in
the Redis crash report. In a long running process eventaully a
significant number of physical memory pages can be tested, increasing
the possibility of early detection of memory errors.

This commit provides an implementation of this idea so that the Redis
community can evaluate how it works in practice.